backend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1.1
django-allauth==65.0.2
django-axes==6.5.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
